✨마이 페이지 컴포넌트 리펙토링
